refactor: admin/user 회원가입 중복 로직을 createSignupPage 함수로 통합
